Split `Query` and `Target` and critical renames in SyncEngine and LocalStore
